# Refactor CTC APIs