… keras/src/backend/tensorflow/nn.py
… keras/src/losses/losses.py
… keras/src/ops/nn.py
… keras/src/ops/nn_test.py