This PR:
- Ensure dtype inference for `ctc_loss` across all backends
- Add `ctc_loss` to numpy backend
- Add `ctc_decode` (including both `"greedy"` and `"beam_search"` strategies) to numpy backend
- Add `"greedy"` strategy for `ctc_decode` to torch backend
- Reduce the number of import lines by using `keras.src.backend`

Furthermore, we can simply utilize `tf.nn.gelu` instead of a custom implementation since we have already removed the usage of `enable_numpy_behavior`