@james77777778 2024-04-25
UPDATES:
- Update the `ctc_loss` implementation in both jax and numpy

I'm not sure about the reference of the original implementation, but I've found that `optax.ctc_loss_with_forward_probs` is much faster. Therefore, I've made modifications based on it.

|master|this PR|
|-|-|
|~4.7s|~2.0s|

- RTX 4070
- CUDA 12.3

Benchmark script:
```python
import time

import jax

from keras import random
from keras.src.backend.jax.nn import ctc_loss

batch_size = 128
steps = 1024
classes = 256
labels = random.randint(shape=[batch_size, steps], minval=1, maxval=classes)
outputs = random.normal(shape=[batch_size, steps, classes])
length = random.randint(shape=[batch_size], minval=steps // 2, maxval=steps - 1)


fn = jax.jit(ctc_loss)

# Warmup
loss = fn(labels, outputs, length, length)
jax.block_until_ready(loss)

# Benchmark
st = time.time()
for _ in range(100):
    loss = fn(labels, outputs, length, length)
    jax.block_until_ready(loss)
ed = time.time()

print(f"{ed-st:.3f}s")

```

cc @MaanasArora