@james77777778 2024-04-25
I've found that the current implementation of `ctc_decode` mismatch with Keras 2:
https://github.com/keras-team/tf-keras/blob/master/tf_keras/backend.py#L7181-L7232

It lacks post-processing of the sparse tensor and doesn't set the blank label as `-1`.

As a result, I've refactored `ctc_decode` in all backends.
Additionally, the jax version of `ctc_decode` is jittable and it should be much faster than before.

I can implement the tf version using low-level ops to support `tf.function(jit_compile=True)` in the next PR.

cc @MaanasArora